fpl_requirements.txt
python-telegram-bot==20.7
aiohttp==3.9.1
sqlite3
python-dotenv==1.0.0
